backend/requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-appconf==1.0.5
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
Markdown==3.3.6
marshmallow==3.14.0
mypy-extensions==0.4.3
oauthlib==3.1.1
Panda3D==1.10.10
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.20
PyJWT==2.3.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
